Refactoring core and accounts app
