+++ accounts/models.py
@@ -1,6 +1,5 @@
 from django.utils.translation import ugettext_lazy as _
-from django.contrib.auth.models import Group, Permission
-from django.shortcuts import get_object_or_404
+from django.contrib.auth.models import Group
 from django.core import validators
 from django.conf import settings
 from django.db import models
