+++ core/email.py
@@ -11,13 +11,15 @@ def send_email_template(subject,
                         from_email=settings.DEFAULT_FROM_EMAIL,
                         fail_silently=False):
     """
-    Template to send email.
+    Function to send email.
     """
 
     # Generate the html template message in string form
     message_html = render_to_string(template, context)
+
     # Generate the text message from html removing html tags
     message_txt = strip_tags(message_html)
+
     # Create the email with multiple alternatives to render html if it can
     email = EmailMultiAlternatives(
         subject=subject,
@@ -25,7 +27,9 @@ def send_email_template(subject,
         from_email=from_email,
         to=recipient_list
     )
+
     # If the email reader accepts the html form as an alternative, use it
     email.attach_alternative(message_html, "text/html")
+
     # When email sending fails, it will silently throw an exception or not
     email.send(fail_silently=fail_silently)
