+++ core/templates/core/news_details.html
@@ -34,7 +34,6 @@
               <br />
             {% else %}
               <p>{{news.title}}</p>
-              <br />
             {% endif %}
             <small>({{ news.created_at }})</small>
           </h2>
